move BC validation to base validator (#2295)
* move BC validation to base validator

* fix unit tests

* add release notes

* fix rn
