+++ demisto_sdk/commands/common/hook_validations/field_base_validator.py
@@ -77,6 +77,7 @@ class FieldBaseValidator(ContentEntityValidator):
 
         is_bc_broke = any(
             [
+                super().is_backward_compatible(),
                 self.is_changed_type(),
                 self.is_changed_from_version(),
             ]
