+++ demisto_sdk/commands/common/hook_validations/incident_type.py
@@ -26,6 +26,7 @@ class IncidentTypeValidator(ContentEntityValidator):
 
         is_bc_broke = any(
             [
+                super().is_backward_compatible(),
                 self.is_changed_from_version()
             ]
         )
