+++ demisto_sdk/commands/common/hook_validations/mapper.py
@@ -58,6 +58,7 @@ class MapperValidator(ContentEntityValidator):
         """Check whether the Mapper is backward compatible or not, update the _is_valid field to determine that"""
 
         answers = [
+            super().is_backward_compatible(),
             self.is_field_mapping_removed(),
         ]
         return not any(answers)
